Start building grid
