+++ src/features/blog/containers/newsContainer/index.js
@@ -4,6 +4,7 @@ import Title from '../../components/title'
 import ImageBanner from '../../../../components/imageBanner'
 import theme from '../../../../theme/theme'
 import NewsCard from '../../components/newsCard'
+import { FlexColumn, StyledFlipMove } from '../../../../components/grid'
 import {
   ALL_ARTICLES,
   NEWS,
@@ -15,8 +16,55 @@ import {
 // import PropTypes from 'prop-types'
 
 const Container = styled.div`
+  display: flex;
+  flex-direction: row;
   margin: 1rem;
 `
+
+const mockBlogs = [
+  {
+    type: ALL_ARTICLES,
+  },
+  {
+    type: NEWS,
+  },
+  {
+    type: RESEARCH,
+  },
+  {
+    type: ANNOUNCEMENTS,
+  },
+  {
+    type: PARTNERS,
+  },
+  {
+    type: RESEARCH,
+  },
+  {
+    type: RESEARCH,
+  },
+  {
+    type: ANNOUNCEMENTS,
+  },
+  {
+    type: PARTNERS,
+  },
+  {
+    type: RESEARCH,
+  },
+  {
+    type: RESEARCH,
+  },
+  {
+    type: ANNOUNCEMENTS,
+  },
+  {
+    type: PARTNERS,
+  },
+  {
+    type: RESEARCH,
+  },
+]
 
 const NewsContainer = () => {
   return (
@@ -30,11 +78,22 @@ const NewsContainer = () => {
         allowContentUnderflow
       />
       <Container>
-        <NewsCard type={ALL_ARTICLES} />
-        <NewsCard type={NEWS} />
-        <NewsCard type={RESEARCH} />
-        <NewsCard type={ANNOUNCEMENTS} />
-        <NewsCard type={PARTNERS} />
+        <StyledFlipMove>
+          {mockBlogs.map(({ type }) => {
+            return (
+              <FlexColumn
+                width={[
+                  1, // 100% between 0px screen width and first breakpoint (375px)
+                  1, // 100% between first breakpoint(375px) and second breakpoint (768px)
+                  1 / 2, // 50% between second breakpoint(768px) and third breakpoint (1024px)
+                  1 / 3, // 33% between third breakpoint(1280px) and fourth breakpoint (1440px)
+                ]}
+              >
+                <NewsCard type={type} />
+              </FlexColumn>
+            )
+          })}
+        </StyledFlipMove>
       </Container>
     </div>
   )
