+++ src/pages/events.js
@@ -1,38 +1,29 @@
 import React, { Fragment, Component } from 'react'
 import PropTypes from 'prop-types'
-import FlipMove from 'react-flip-move'
 import styled from 'styled-components'
 import moment from 'moment'
 import noScroll from 'no-scroll'
 import { media } from '../theme/media'
 import theme from '../theme/theme'
-import { EventListingCard } from '../features/events'
-import EventsFilters from '../features/events/components/eventsFilters'
+import { EventListingCard } from '../features/Events'
+import EventsFilters from '../features/Events/components/eventsFilters'
 import ImageBanner from '../components/imageBanner'
 import Button from '../components/button'
-import { Container, Row, Column } from '../components/grid'
+import {
+  Container,
+  Row,
+  Column,
+  StyledFlipMove,
+  FlexColumn,
+} from '../components/grid'
 import { Consumer } from '../components/appContext'
-import { filterByLimit } from '../features/events/helpers'
+import { filterByLimit } from '../features/Events/helpers'
 import { dateFormat } from '../constants'
 import filterIcon from '../theme/assets/images/icon-filters.svg'
-
-const FlexColumn = styled(Column)`
-  display: block;
-
-  ${media.tablet`
-    display: flex;
-  `};
-`
 
 const ColumnTextCenter = styled(Column)`
   text-align: center;
 `
-const StyledFlipMove = styled(FlipMove)`
-  display: flex;
-  flex-wrap: wrap;
-  flex-basis: 100%;
-`
-
 const ContainerAddFilters = styled(Container)`
   padding: 20px 0;
   margin-bottom: 20px;
